refactor(dsl-mesh): split tessellate out of cleanup; drop dead triangle-domain ops
Two semantically separate concerns lived together in csg::cleanup:
fixing the polygon stream (welding, coplanar merging, T-junction
repair, sliver removal) and converting that polygon mesh into
triangles for the GPU. They share a common input shape but answer
different questions and have different consumers — under ADR-0057
the polygon-domain public API skips tessellation entirely.

Move CDT triangulation into a new csg::tessellate sibling module:

- crates/aether-dsl-mesh/src/csg/tessellate.rs — owns triangulate_indexed
  (the indexed-mesh→triangle-polygon conversion that was
  cleanup::mesh::IndexedMesh::cdt_triangulate), tessellate_polygon_f32
  (display-time entry, was cleanup::tessellate_polygon_f32), and run
  (cleanup + CDT for the wire path).
- crates/aether-dsl-mesh/src/csg/tessellate/cdt — moved from
  cleanup/cdt; the CDT implementation has no dependency on the
  cleanup passes themselves, only on the indexed mesh shape.
- crates/aether-dsl-mesh/src/csg/cleanup.rs — keeps run_to_loops
  (n-gon path) plus a new run_to_indexed (indexed-mesh path that
  tessellate consumes without round-tripping through Vec<Polygon>).
  The legacy run is gone; mesh::mesh now calls csg::tessellate::run
  directly.
- IndexedMesh / IndexedPolygon / VertexId visibility: bumped from
  pub(super) to pub(in crate::csg) so tessellate can name them.

Also drop the dead pre-ADR-0057 triangle-domain layer:

- csg::ops::{union, intersection, difference} previously returned
  triangulated Vec<Polygon> by calling cleanup::run; the
  parallel _polygons variants returned n-gon loops via run_to_loops.
  The mesh path uses _raw + cleanup::run_to_loops, never the
  triangulated entries — they were zero-caller dead code in the
  whole workspace. Collapse: rename _polygons → no suffix (canonical
  names), delete the triangle-returning variants.
- csg::{union_triangles, intersection_triangles, difference_triangles}
  and triangles_to_polygons: same story, zero callers anywhere,
  deleted.

Triangulation now happens at exactly one place (csg::tessellate::run)
called by exactly one caller (mesh::mesh) for the wire Vec<Triangle>
path; the polygon path (mesh::mesh_polygons → polygon::mesh_polygons)
never touches it. Behavior unchanged: 235 lib tests + 24 regression
tests pass, 1 ignored (the lathe BSP perf cliff, issue 300).
