+++ crates/aether-dsl-mesh/src/csg/cleanup.rs
@@ -1,11 +1,10 @@
 //! Post-CSG mesh cleanup pipeline (ADR-0055, refactored under ADR-0057).
 //!
 //! Runs on the polygon stream produced by `ops::{union, intersection,
-//! difference}` before triangulation back to the wire `Vec<Triangle>`
-//! format. The pipeline operates in the same fixed-point integer domain
-//! as the BSP CSG core; passes are pure and exact.
+//! difference}`. The pipeline operates in the same fixed-point integer
+//! domain as the BSP CSG core; passes are pure and exact.
 //!
-//! Five passes (composed in order):
+//! Four passes (composed in order):
 //!
 //! 1. **Vertex welding** — converts owned-vertex polygons into an
 //!    indexed-mesh representation, deduplicating vertices by exact
@@ -24,38 +23,35 @@
 //!    drifting beyond the welding tolerance). Edge-triggered, not
 //!    coordinate-triggered, so it doesn't risk colliding distinct
 //!    features.
-//! 5. **CDT triangulation for the wire** — groups loops by plane,
-//!    runs constrained Delaunay triangulation per group (ADR-0056),
-//!    and emits triangle polygons. Multi-loop groups (faces with
-//!    holes) are triangulated as a single CDT call so the hole is
-//!    cut out cleanly.
 //!
-//! Pass 5 is what makes [`run`] return the same `Vec<Polygon>` shape
-//! `csg::ops` expects today. The polygon-domain entry point
-//! [`run_to_loops`] stops after pass 4 and returns n-gons.
+//! Triangulation (CDT) is the responsibility of [`super::tessellate`],
+//! not cleanup. The polygon-domain public API ([`run_to_loops`] +
+//! [`crate::polygon::mesh_polygons`]) skips tessellation entirely
+//! because n-gon polygons are the canonical mesh form per ADR-0057;
+//! the legacy triangle-domain ops in [`super::ops`] compose cleanup
+//! and tessellation explicitly via [`super::tessellate::run`].
 
-mod cdt;
 mod merge;
-mod mesh;
+pub(in crate::csg) mod mesh;
 mod slivers;
 mod tjunctions;
 mod weld;
 
-use crate::csg::plane::Plane3;
-use crate::csg::point::Point3;
 use crate::csg::polygon::Polygon;
 
-/// Run the cleanup pipeline on a polygon list, returning triangulated
-/// polygons (3 vertices each) for the wire `Vec<Triangle>` path.
-pub fn run(polygons: Vec<Polygon>) -> Vec<Polygon> {
+/// Run the cleanup pipeline and return the final indexed mesh, ready
+/// for either polygon-domain consumption ([`mesh::IndexedMesh::into_polygons`])
+/// or triangulation ([`super::tessellate::run`]). Internal entry point
+/// that lets `tessellate` reuse the indexed representation without
+/// round-tripping through `Vec<Polygon>`.
+pub(in crate::csg) fn run_to_indexed(polygons: Vec<Polygon>) -> mesh::IndexedMesh {
     mesh::IndexedMesh::weld(polygons)
         .merge_coplanar()
         .repair_tjunctions()
         .remove_slivers()
-        .cdt_triangulate()
 }
 
-/// Run the cleanup pipeline on a polygon list, stopping after pass 3 so
+/// Run the cleanup pipeline on a polygon list, stopping after pass 4 so
 /// the output stays in n-gon-loop form (one polygon per boundary loop).
 /// This is the entry point for the polygon-domain public API per
 /// ADR-0057 — `mesh_polygons` calls this and groups loops into outer +
@@ -66,151 +62,5 @@ pub fn run(polygons: Vec<Polygon>) -> Vec<Polygon> {
 /// plus the CW hole loop. Callers responsible for grouping by plane
 /// when they want the polygon-with-holes shape.
 pub fn run_to_loops(polygons: Vec<Polygon>) -> Vec<Polygon> {
-    mesh::IndexedMesh::weld(polygons)
-        .merge_coplanar()
-        .repair_tjunctions()
-        .remove_slivers()
-        .into_polygons()
-}
-
-/// Display-time tessellation for the polygon-domain public API
-/// (ADR-0057). Takes a polygon-with-holes in f32 coords, runs the
-/// internal CDT against the integer fixed-point pool, and returns
-/// triangles in f32 for GPU upload.
-///
-/// `outer` is the CCW outer boundary; `holes` are CW inner boundaries.
-/// Returns `None` if the inputs collapse to fewer than 3 unique
-/// vertices, fall outside the integer fixed-point coordinate budget
-/// (ADR-0054 ±256 unit cap), or CDT fails to enforce a constraint.
-///
-/// Callers should fall back to fan triangulation on `None` so geometry
-/// isn't dropped silently.
-pub fn tessellate_polygon_f32(
-    outer: &[[f32; 3]],
-    holes: &[Vec<[f32; 3]>],
-) -> Option<Vec<[[f32; 3]; 3]>> {
-    if outer.len() < 3 {
-        return None;
-    }
-
-    // Convert to integer fixed-point and build a flat vertex pool.
-    let mut vertices: Vec<Point3> =
-        Vec::with_capacity(outer.len() + holes.iter().map(|h| h.len()).sum::<usize>());
-    let mut outer_indices: Vec<usize> = Vec::with_capacity(outer.len());
-    for v in outer {
-        let p = Point3::from_f32(*v).ok()?;
-        outer_indices.push(vertices.len());
-        vertices.push(p);
-    }
-    let mut hole_index_loops: Vec<Vec<usize>> = Vec::with_capacity(holes.len());
-    for hole in holes {
-        let mut indices = Vec::with_capacity(hole.len());
-        for v in hole {
-            let p = Point3::from_f32(*v).ok()?;
-            indices.push(vertices.len());
-            vertices.push(p);
-        }
-        hole_index_loops.push(indices);
-    }
-
-    // Compute the plane from the outer loop's first three integer
-    // vertices. The CDT uses this for axis selection only; the CCW
-    // outer assumption gives a normal pointing "outward" by construction.
-    if outer_indices.len() < 3 {
-        return None;
-    }
-    let plane = Plane3::from_points(
-        vertices[outer_indices[0]],
-        vertices[outer_indices[1]],
-        vertices[outer_indices[2]],
-    );
-    if plane.is_degenerate() {
-        return None;
-    }
-
-    let mut all_loops: Vec<Vec<usize>> = Vec::with_capacity(1 + holes.len());
-    all_loops.push(outer_indices);
-    all_loops.extend(hole_index_loops);
-
-    let triangles = cdt::triangulate_loops(&vertices, &all_loops, &plane)?;
-    Some(
-        triangles
-            .into_iter()
-            .map(|tri| {
-                [
-                    vertices[tri[0]].to_f32(),
-                    vertices[tri[1]].to_f32(),
-                    vertices[tri[2]].to_f32(),
-                ]
-            })
-            .collect(),
-    )
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn tessellate_polygon_f32_rejects_fewer_than_3_outer_vertices() {
-        let outer = vec![[0.0, 0.0, 0.0], [1.0, 0.0, 0.0]];
-        let result = tessellate_polygon_f32(&outer, &[]);
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn tessellate_polygon_f32_rejects_out_of_range_coordinates() {
-        // ADR-0054 caps fixed-point coordinates at ±256. Larger values
-        // must fail-fast at the f32→fixed conversion rather than silently
-        // truncate.
-        let outer = vec![
-            [0.0, 0.0, 0.0],
-            [300.0, 0.0, 0.0], // out of range
-            [0.0, 1.0, 0.0],
-        ];
-        let result = tessellate_polygon_f32(&outer, &[]);
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn tessellate_polygon_f32_rejects_degenerate_collinear_outer() {
-        // Three collinear points → degenerate plane → None.
-        let outer = vec![[0.0, 0.0, 0.0], [1.0, 0.0, 0.0], [2.0, 0.0, 0.0]];
-        let result = tessellate_polygon_f32(&outer, &[]);
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn tessellate_polygon_f32_happy_path_square_with_hole() {
-        let outer = vec![
-            [0.0, 0.0, 0.0],
-            [4.0, 0.0, 0.0],
-            [4.0, 4.0, 0.0],
-            [0.0, 4.0, 0.0],
-        ];
-        let hole = vec![
-            [1.0, 1.0, 0.0],
-            [1.0, 3.0, 0.0],
-            [3.0, 3.0, 0.0],
-            [3.0, 1.0, 0.0],
-        ];
-        let tris = tessellate_polygon_f32(&outer, &[hole]).expect("annular should triangulate");
-        // Topological minimum for 8-vertex annular = 8 triangles.
-        assert_eq!(tris.len(), 8);
-        // Total area = outer (16) - hole (4) = 12. Use shoelace on each
-        // triangle (signed) and sum.
-        let signed_double_area: f32 = tris
-            .iter()
-            .map(|tri| {
-                (tri[1][0] - tri[0][0]) * (tri[2][1] - tri[0][1])
-                    - (tri[1][1] - tri[0][1]) * (tri[2][0] - tri[0][0])
-            })
-            .sum();
-        // Doubled area of annular region = 24. Allow a small tolerance
-        // for f32 rounding through the integer fixed-point round trip.
-        assert!(
-            (signed_double_area - 24.0).abs() < 0.01,
-            "annular doubled area mismatch: {signed_double_area}"
-        );
-    }
+    run_to_indexed(polygons).into_polygons()
 }
