+++ crates/aether-dsl-mesh/src/csg/ops.rs
@@ -11,23 +11,15 @@
 //! Inputs and outputs are flat polygon lists — this module does no I/O,
 //! no parsing, no rendering.
 //!
-//! ### Output shape: triangulated vs n-gon loops
+//! ### Output shape
 //!
-//! Each operation has two public entry points that differ only in the
-//! cleanup pass run after BSP composition:
-//!
-//! - [`union`] / [`intersection`] / [`difference`] — output is fully
-//!   triangulated (3-vertex polygons, CDT-routed). Suitable for the
-//!   wire `Vec<Triangle>` path.
-//! - [`union_polygons`] / [`intersection_polygons`] /
-//!   [`difference_polygons`] — output is n-gon boundary loops, no CDT.
-//!   Suitable for the polygon-domain path that triangulates only at
-//!   the wire/render boundary (avoids the `Polygon::from_triangle`
-//!   plane re-derivation step that flips sliver normals after f32
-//!   round-tripping).
-//!
-//! BSP composition is shared between the two via the `_raw` helpers
-//! ([`union_raw`], [`intersection_raw`], [`difference_raw`]).
+//! [`union`] / [`intersection`] / [`difference`] return n-gon boundary
+//! loops (no triangulation) — the canonical mesh form per ADR-0057.
+//! BSP composition is shared with the no-cleanup `_raw` helpers
+//! ([`union_raw`], [`intersection_raw`], [`difference_raw`]) which the
+//! AST mesh evaluator uses to chain CSG operations without inter-op
+//! cleanup. The wire `Vec<Triangle>` path triangulates only at the
+//! root via [`super::tessellate::run`].
 
 use super::CsgError;
 use super::bsp::BspTree;
@@ -35,31 +27,14 @@ use super::cleanup;
 use super::polygon::Polygon;
 
 pub fn union(a: Vec<Polygon>, b: Vec<Polygon>) -> Result<Vec<Polygon>, CsgError> {
-    Ok(cleanup::run(union_raw(a, b)?))
-}
-
-pub fn intersection(a: Vec<Polygon>, b: Vec<Polygon>) -> Result<Vec<Polygon>, CsgError> {
-    Ok(cleanup::run(intersection_raw(a, b)?))
-}
-
-pub fn difference(a: Vec<Polygon>, b: Vec<Polygon>) -> Result<Vec<Polygon>, CsgError> {
-    Ok(cleanup::run(difference_raw(a, b)?))
-}
-
-/// Polygon-domain `union` — same BSP composition as [`union`], but
-/// cleanup stops at n-gon boundary loops instead of CDT-triangulating
-/// for the wire.
-pub fn union_polygons(a: Vec<Polygon>, b: Vec<Polygon>) -> Result<Vec<Polygon>, CsgError> {
     Ok(cleanup::run_to_loops(union_raw(a, b)?))
 }
 
-/// Polygon-domain `intersection` — see [`union_polygons`].
-pub fn intersection_polygons(a: Vec<Polygon>, b: Vec<Polygon>) -> Result<Vec<Polygon>, CsgError> {
+pub fn intersection(a: Vec<Polygon>, b: Vec<Polygon>) -> Result<Vec<Polygon>, CsgError> {
     Ok(cleanup::run_to_loops(intersection_raw(a, b)?))
 }
 
-/// Polygon-domain `difference` — see [`union_polygons`].
-pub fn difference_polygons(a: Vec<Polygon>, b: Vec<Polygon>) -> Result<Vec<Polygon>, CsgError> {
+pub fn difference(a: Vec<Polygon>, b: Vec<Polygon>) -> Result<Vec<Polygon>, CsgError> {
     Ok(cleanup::run_to_loops(difference_raw(a, b)?))
 }
 
@@ -339,9 +314,9 @@ mod tests {
     }
 
     /// **Diagnostic**: compares boundary-edge count of `box - sphere`
-    /// before vs after `cleanup::run`. Localizes whether the regression's
-    /// 36 boundary edges originate in the BSP composition or in the
-    /// post-BSP cleanup pipeline.
+    /// before vs after the cleanup pipeline. Localizes whether the
+    /// regression's 36 boundary edges originate in the BSP composition
+    /// or in the post-BSP cleanup pipeline.
     ///
     /// Reads as: print both numbers via panic so the result is captured
     /// in test output. Will be deleted or converted to an assertion
@@ -1032,80 +1007,37 @@ mod tests {
         }
     }
 
-    /// `union_polygons` returns n-gon loops, not the per-triangle
-    /// polygons `union` returns. For a `2-box ∪ 2-box overlap` the loop
-    /// form has fewer polygons (one per face component) and at least
-    /// one polygon with >3 vertices (the merged face quads).
+    /// `union` returns n-gon boundary loops aggregated by coplanar
+    /// merging — for a `2-box ∪ 2-box overlap` the result has at least
+    /// one polygon with >3 vertices (the merged face quads). Verifies
+    /// the cleanup pipeline composes through to the public entry.
     #[test]
-    fn union_polygons_returns_n_gon_loops() {
+    fn union_returns_n_gon_loops() {
         let a = axis_aligned_box(0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 0);
         let b = axis_aligned_box(0.5, 0.0, 0.0, 1.0, 1.0, 1.0, 1);
-        let triangulated = union(a.clone(), b.clone()).unwrap();
-        let loops = union_polygons(a, b).unwrap();
-        assert!(
-            loops.len() < triangulated.len(),
-            "loops should aggregate triangles"
-        );
+        let loops = union(a, b).unwrap();
         assert!(
             loops.iter().any(|p| p.vertices.len() > 3),
             "expected at least one n-gon loop with >3 vertices"
         );
-        assert!(triangulated.iter().all(|p| p.vertices.len() == 3));
     }
 
     #[test]
-    fn intersection_polygons_returns_n_gon_loops() {
+    fn intersection_returns_non_empty_for_overlapping_inputs() {
         let a = axis_aligned_box(0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 0);
         let b = axis_aligned_box(0.5, 0.5, 0.5, 0.7, 0.7, 0.7, 1);
-        let triangulated = intersection(a.clone(), b.clone()).unwrap();
-        let loops = intersection_polygons(a, b).unwrap();
+        let loops = intersection(a, b).unwrap();
         assert!(
             !loops.is_empty(),
             "overlapping intersection must be non-empty"
         );
-        assert!(loops.len() < triangulated.len());
-        assert!(triangulated.iter().all(|p| p.vertices.len() == 3));
     }
 
     #[test]
-    fn difference_polygons_returns_n_gon_loops() {
+    fn difference_returns_non_empty_for_partial_subtractor() {
         let outer = axis_aligned_box(0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 0);
         let cutter = axis_aligned_box(0.0, 0.0, 0.0, 0.5, 0.5, 0.5, 1);
-        let triangulated = difference(outer.clone(), cutter.clone()).unwrap();
-        let loops = difference_polygons(outer, cutter).unwrap();
+        let loops = difference(outer, cutter).unwrap();
         assert!(!loops.is_empty());
-        assert!(loops.len() < triangulated.len());
-        assert!(triangulated.iter().all(|p| p.vertices.len() == 3));
-    }
-
-    /// The polygon-domain entry must agree with the triangulated entry
-    /// on the underlying surface — same colors present, same plane
-    /// directions. Verifies the cleanup-tail swap doesn't change what
-    /// BSP composed.
-    #[test]
-    fn polygon_and_triangulated_entries_agree_on_surface_topology() {
-        let a = axis_aligned_box(0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 0);
-        let b = axis_aligned_box(0.0, 0.0, 0.0, 0.5, 0.5, 0.5, 1);
-        let tris = difference(a.clone(), b.clone()).unwrap();
-        let loops = difference_polygons(a, b).unwrap();
-
-        let collect_colors = |polys: &[Polygon]| -> std::collections::BTreeSet<u32> {
-            polys.iter().map(|p| p.color).collect()
-        };
-        assert_eq!(collect_colors(&tris), collect_colors(&loops));
-
-        let collect_plane_signs = |polys: &[Polygon]| -> std::collections::BTreeSet<(i8, i8, i8)> {
-            polys
-                .iter()
-                .map(|p| {
-                    (
-                        p.plane.n_x.signum() as i8,
-                        p.plane.n_y.signum() as i8,
-                        p.plane.n_z.signum() as i8,
-                    )
-                })
-                .collect()
-        };
-        assert_eq!(collect_plane_signs(&tris), collect_plane_signs(&loops));
     }
 }
