+++ crates/aether-dsl-mesh/src/csg/tessellate/cdt.rs
@@ -1,19 +1,17 @@
-//! Constrained Delaunay triangulation for the cleanup pipeline (ADR-0056).
+//! Constrained Delaunay triangulation for the tessellation pass (ADR-0056).
 //!
 //! Replaces the ear-clipping + hole-bridging step that ADR-0055 ships;
 //! produces sliver-free, locally-Delaunay triangulations of polygons-
 //! with-holes by enforcing boundary edges as constraints rather than
 //! splicing holes into the outer loop with a slit.
 //!
-//! Module layout (built up across the implementation cascade):
+//! Module layout:
 //!
 //! - [`predicates`]: exact integer in-circle and orient2d tests in
 //!   i128. Foundation for everything else.
-//! - `bowyer_watson`: incremental Delaunay triangulation (PR 2, not
-//!   yet shipped).
-//! - `triangulate`: constraint enforcement + inside/outside marking +
-//!   the public entry point that `merge::process_component` will call
-//!   (PR 3, not yet shipped).
+//! - [`bowyer_watson`]: incremental Delaunay triangulation.
+//! - [`triangulate`]: constraint enforcement + inside/outside marking
+//!   + the public entry point that `triangulate_indexed` calls.
 
 pub(super) mod bowyer_watson;
 pub(super) mod predicates;
