+++ crates/aether-dsl-mesh/src/csg/tessellate/cdt/predicates.rs
